refactor: split handler service boundaries
