+++ package.json
@@ -71,6 +71,7 @@
     "@neondatabase/api-client": "^2.7.1",
     "@neondatabase/serverless": "^1.0.1",
     "@rollup/plugin-commonjs": "^28.0.3",
+    "@sentry/electron": "^7.13.0",
     "@tailwindcss/typography": "^0.5.16",
     "@tailwindcss/vite": "^4.1.3",
     "@tanstack/react-query": "^5.75.5",
